Structured the view models a bit
