+++ BlazorQuiz/Shared/ViewModels/CreateQuizDto.cs
@@ -1,10 +1,28 @@
-﻿namespace BlazorQuiz.Shared.ViewModels
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace BlazorQuiz.Shared.ViewModels
 {
+
+    public class QuestionSharedViewModel
+    {
+        public string Question { get; set; }
+        public string Answer1 { get; set; }
+        public string Answer2 { get; set; }
+        public string Answer3 { get; set; }
+        public string Answer4 { get; set; }
+        public string? QuizMediaUrl { get; set; } = string.Empty;
+        public int Id { get; set; }
+    }
+
     public class NewQuestionViewModel
     {
         public NewQuestionViewModel()
         {
-            
+
         }
         public NewQuestionViewModel(NewQuestionViewModel clone)
         {
@@ -14,9 +32,10 @@
             Answer3 = clone.Answer3;
             Answer4 = clone.Answer4;
             CorrectAnswer = clone.Answer1;
-           
+
         }
         public string QuizImageUrl { get; set; } = "";
+        public string QuizImageFullPath { get; set; } = "";
         public string Question { get; set; }
         public string Answer1 { get; set; }
         public string Answer2 { get; set; }
@@ -32,7 +51,16 @@
             Answer3 = string.Empty;
             Answer4 = string.Empty;
             QuizImageUrl = string.Empty;
+            QuizImageFullPath = string.Empty;
         }
 
+    }
+
+    public class NewQuizViewModel
+    {
+        public string Title { get; set; }
+        public string? PublicId { get; set; }
+        public int Timer { get; set; }
+        public List<NewQuestionViewModel> Questions { get; set; }
     }
 }
