--- BlazorQuiz/Shared/ViewModels/GuessCheckViewModel.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace BlazorQuiz.Shared.ViewModels
-{
-    public class GuessCheckViewModel
-    {
-        public int GuessId { get; set; }
-        public string Guess { get; set;}
-        public int Seconds { get; set;}
-
-    }
-}
