--- BlazorQuiz/Shared/ViewModels/QuestionSharedViewModel.cs
@@ -1,13 +1,0 @@
-﻿namespace BlazorQuiz.Shared.ViewModels
-{
-    public class QuestionSharedViewModel
-    {
-        public string Question { get; set; }
-        public string Answer1 { get; set; }
-        public string Answer2 { get; set; }
-        public string Answer3 { get; set; }
-        public string Answer4 { get; set; }
-        public string? QuizMediaUrl { get; set; } = string.Empty;
-        public int Id { get; set; }
-    }
-}
