--- BlazorQuiz/Shared/ViewModels/UserCreatedQuizViewModel.cs
@@ -1,8 +1,0 @@
-﻿namespace BlazorQuiz.Shared.ViewModels
-{
-    public class UserCreatedQuizViewModel
-    {
-        public string PublicId { get; set; }
-        public string Name { get; set; }
-    }
-}
