--- BlazorQuiz/Shared/ViewModels/UserQuizViewModel.cs
@@ -1,8 +1,0 @@
-﻿namespace BlazorQuiz.Shared.ViewModels
-{
-    public class UserQuizViewModel
-    {
-        public int Score { get; set; }
-        public string User { get; set; }
-    }
-}
